move version check to cli.js
